rename evaluate dm
